add support for options and fix failing test
Not passing options to autoprefixer changes semantics over time, as it's
default fallback value is something like the latest 2 versions of
browsers.

While adding support for passing options through to autoprefixer I
fixated the version which the basic test uses, and added another test
that asserts that the cascade option you can pass to autoprefixer is
also acknowledged.